requirements.txt
pytorch_lightning==1.9.0
Pillow==10.0.0
opencv-python
transformers
taming-transformers-rom1504 
tqdm
numpy
kornia
webdataset
omegaconf
einops
scikit-image
pymcubes
carvekit-colab
open3d
trimesh
easydict
nerfacc
transforms3d
imageio-ffmpeg==0.4.7
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/openai/CLIP.git
